// src/interfaces/user.interfaces.ts
export interface UserInterface {
    _id: string
    firstName: string
    lastName: string
    username: string
    email: string
    password: string
    posts: string[]
    comments: string[]
    chats: string[]
    followers: string[]
    following: string[]
    createdAt: string
    updatedAt: string
}